requirements.txt
fastapi
uvicorn[standard]
pyppeteer
python-dotenv